requirements.txt
torch
einops
numpy
torchtext==0.6.0
tqdm
matplotlib
seaborn
six
pandas
plotly-express